Updated SelectCell to PickerCell
